backend/requirements.txt
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
Flask==2.3.3
Flask-Cors==4.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.32.3
urllib3==2.4.0
Werkzeug==2.3.7
